Reverted idb PersistApi and moved it into database
