+++ app/assets/js/src/database/index.ts
@@ -1,1 +1,1 @@
-export { adapterV1 } from './adapterV1';
+export { adapterV1, dbV1 } from './adapterV1';
