+++ app/assets/js/src/persist/idb.ts
@@ -1,17 +1,10 @@
-import { StorageKey } from 'data/shared';
 import type { PersistApi } from 'persist/PersistApi';
-import {
-	doDatabaseTransaction,
-} from 'utils';
-import { getDatabase } from 'utils/indexedDB';
+import { doDatabaseTransaction } from 'utils';
 
-import { adapterV1 } from 'database';
-import { getDatabaseVersion } from 'database/utils';
 import { ObjectStoreName } from 'database/metadata';
 
 /**
- * A {@linkcode PersistApi} interface for working with the
- * IndexedDB API.
+ * A {@linkcode PersistApi} interface for working with the IndexedDB API, using the database schema v1.
  */
 export const idb: PersistApi = {
 	async set(key, data) {
@@ -22,30 +15,12 @@ export const idb: PersistApi = {
 		);
 	},
 
-	async get(key) {
-		// TODO: Get rid of this unnecessary retrieval, just used to make sure the database has been upgraded first
-		await getDatabase();
-		// TODO: Get rid of this database v1 handling
-		const dbVersion = await getDatabaseVersion();
-		if (dbVersion === 1 || dbVersion === null) {
-			return doDatabaseTransaction(
-				'readonly',
-				[ObjectStoreName.DATA],
-				([objectStore]) => objectStore.get(key)
-			);
-		}
-
-		if (key === StorageKey.DAYS) {
-			return adapterV1.getDays();
-		} else if (key === StorageKey.DAY_TASKS) {
-			return adapterV1.getDayTasks();
-		} else if (key === StorageKey.TASKS) {
-			return adapterV1.getTasks();
-		} else if (key === StorageKey.TEMPLATES) {
-			return adapterV1.getTemplates();
-		} else {
-			throw new RangeError(`Unrecognised key ${key}`);
-		}
+	get(key) {
+		return doDatabaseTransaction(
+			'readonly',
+			[ObjectStoreName.DATA],
+			([objectStore]) => objectStore.get(key)
+		);
 	},
 
 	async delete(key) {
